homework p2 jdbc
