+++ Lesson22jdbc/src/main/java/tms/homework/utils/CRUDUtils.java
@@ -3,36 +3,34 @@ package tms.homework.utils;
 import tms.homework.model.Worker;
 
 import java.sql.Connection;
-import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 public class CRUDUtils {
-    private final static String DB_URL = "jdbc:postgresql://localhost:5432/c18-onl2022";
-    private final static String DB_LOGIN = "postgres";
-    private final static String DB_PASSWORD = "Barsukov28071992";
-
-    public CRUDUtils() {
-    }
-
-    public static Connection getConnection() {
-        Connection connection = null;
-        try {
-            connection = DriverManager.getConnection(DB_URL, DB_LOGIN, DB_PASSWORD);
-
-        } catch (SQLException e) {
-            System.out.println(e.getMessage() + "Error");
-        }
-        return connection;
-    }
+    private static final String GET_ALL_WORKERS_QUERY = "SELECT * FROM workers_db.workers";
+    private static final String INSERT_WORKERS_QUERY = "INSERT INTO workers_db.workers (id, login, salary, age, date, description) VALUES(?, ?, ?, ?, ?, ?)";
+    private static final String UPDATE_WORKERS_QUERY = "UPDATE workers_db.workers SET course = ? WHERE id = ?";
+    private static final String DELETE_WORKERS_QUERY = "DELETE FROM workers_db.workers WHERE id = ?";
 
     public static List<Worker> getAllWorkers() {
-        List<Worker> workers = new ArrayList<>();
-        try (Connection connection =) {
+        List<Worker> workerList = new ArrayList<>();
+        try (Connection connection = DbUtils.getConnection()) {
+            PreparedStatement preparedStatement = connection.prepareStatement(GET_ALL_WORKERS_QUERY);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                int id = resultSet.getInt("id");
+                String login = resultSet.getString("login");
+                double salary = resultSet.getDouble("salary");
+                Date date = resultSet.getDate("date");
 
+            }
         } catch (SQLException e) {
             System.out.println(e.getMessage() + " Error");
         }
+        return workerList;
     }
 }
